requirement.txt
# Common
Flask==1.0.2                # 基础框架

Flask-SQLAlchemy==2.3.2     # 数据库工具
Flask-WTF==0.14.2           # 表单库
Flask-CKEditor==0.4.3       # 富文本编辑器
Flask-Login==0.4.1          # 用户登陆相关
Bootstrap-Flask==1.0.10     # Bootstrap 前端封装库
Flask-Moment==0.7.0         # timestamp 格式化库

python-dotenv==0.10.1       # 读取 .env .flaskenv 环境配置文件

# Development
watchdog==0.9.0             # 替换 Werkzeug 默认的 stat 重载器
faker==1.0.4                # 用于测试数据生成